# [FIX] K-means slowness