##### Issue
After Orange 3.21 we introduced quite a few performance-regressions to K-means, which became very slow for large data set. 

##### Description of changes

1. Preprocess data once. In master, data set is preprocessed separately for each number of clusters and then also when computing silhouettes. If used with from-to this caused too many in-memory copies of data. Which means crashes on big data due to memory usage. Note that sklearn's k-means makes another copy of the data.

2. Only compute approximate (sampled) silhouettes for big data. Fixes bug introduced in Orange 3.22. Compute them in worker threads.

3. Fix O(|features|^2) when creating centroids.

There is a further obvious improvement that I did not tackle in this PR: preprocessing in a worker thread. Now (=this branch, master, current release) the preprocessing for big data blocks UI for about a minute for a data set with 98304 rows and 806 columns. It is mainly normalization. If it is disabled, it only blocks for a few seconds. 

Benchmarks
=========

Orange  3.21

```
[from_to_100_100] with 3 loops, best of 3:
	min 372 msec per loop
	avg 374 msec per loop
[from_to_100_100_no_normalize] with 3 loops, best of 3:
	min 379 msec per loop
	avg 414 msec per loop
[from_to_sampled_silhouette] with 3 loops, best of 3:
	min 2.1 sec per loop
	avg 2.18 sec per loop
[wide] with 3 loops, best of 3:
	min 306 msec per loop
	avg 315 msec per loop
```

Master

```
[from_to_100_100] with 3 loops, best of 3:
	min 1.98 sec per loop
	avg 1.99 sec per loop
[from_to_100_100_no_normalize] with 3 loops, best of 3:
	min 824 msec per loop
	avg 828 msec per loop
[from_to_sampled_silhouette] with 3 loops, best of 3:
	min 10.6 sec per loop
	avg 10.7 sec per loop
[wide] with 3 loops, best of 3:
	min 2.43 sec per loop
	avg 2.44 sec per loop
```

This branch

```
[from_to_100_100] with 3 loops, best of 3:
	min 636 msec per loop
	avg 663 msec per loop
[from_to_100_100_no_normalize] with 3 loops, best of 3:
	min 462 msec per loop
	avg 486 msec per loop
[from_to_sampled_silhouette] with 3 loops, best of 3:
	min 2.62 sec per loop
	avg 2.99 sec per loop
[wide] with 3 loops, best of 3:
	min 640 msec per loop
	avg 649 msec per loop
```

Still not quite Orange 3.21 performance, but will do. The different in [wide] is due to normalization.


##### Includes
- [X] Code changes
- [X] Benchmarks
- [ ] Documentation